+++ backend/migrations/000015_alter_communities.up.sql
@@ -1,4 +1,4 @@
 ALTER TABLE communities ADD COLUMN contract_name VARCHAR(150);
 ALTER TABLE communities ADD COLUMN contract_addr VARCHAR(150);
 ALTER TABLE communities ADD COLUMN public_path VARCHAR(150);
-ALTER TABLE communities ADD COLUMN threshold INT;
+ALTER TABLE communities ADD COLUMN threshold NUMERIC(18,0);
